+++ tests/solidity_test_helpers/MockShareTokenFactory.sol
@@ -8,19 +8,19 @@ import 'IController.sol';
 
 contract MockShareTokenFactory {
     IMarket private createShareTokenMarketValue;
-    uint8 private createShareTokenOutcomeValue;
-    uint8 private createShareTokenCounter;
+    uint256 private createShareTokenOutcomeValue;
+    uint256 private createShareTokenCounter;
     IShareToken[] private setCreateShareTokenValue;
 
     function resetCreateShareToken() public {
         createShareTokenCounter = 0;
     }
 
-    function getCreateShareTokenCounter() public returns(uint8) {
+    function getCreateShareTokenCounter() public returns(uint256) {
         return createShareTokenCounter;
     }
 
-    function getCreateShareToken(uint8 _index) public returns(IShareToken) {
+    function getCreateShareToken(uint256 _index) public returns(IShareToken) {
         return setCreateShareTokenValue[_index];
     }
 
@@ -32,11 +32,11 @@ contract MockShareTokenFactory {
         return createShareTokenMarketValue;
     }
 
-    function getCreateShareTokenOutcomeValue() public returns(uint8) {
+    function getCreateShareTokenOutcomeValue() public returns(uint256) {
         return createShareTokenOutcomeValue;
     }
-    
-    function createShareToken(IController _controller, IMarket _market, uint8 _outcome) public returns (IShareToken) {
+
+    function createShareToken(IController _controller, IMarket _market, uint256 _outcome) public returns (IShareToken) {
         createShareTokenMarketValue = _market;
         createShareTokenOutcomeValue = _outcome;
         createShareTokenCounter += 1;
